# prototype-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments/{id}": {
      "get": {
        "operationId": "getComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Get a Comment by ID.",
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Update a Comment.",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted."
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Delete a Comment.",
        "tags": [
          "comments"
        ]
      }
    },
...
    "/review-comments": {
      "post": {
        "operationId": "createReviewComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a ReviewComment.",
        "tags": [
          "review-comments"
        ]
      },
      "get": {
        "operationId": "listReviewComments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewCommentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List ReviewComment records.",
        "tags": [
          "review-comments"
        ]
      }
    },
    "/review-comments/{id}": {
      "get": {
        "operationId": "getReviewComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ReviewComment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ReviewComment not found."
          }
        },
        "summary": "Get a ReviewComment by ID.",
        "tags": [
          "review-comments"
        ]
      },
      "patch": {
        "operationId": "updateReviewComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ReviewComment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ReviewComment not found."
          }
        },
        "summary": "Update a ReviewComment.",
        "tags": [
          "review-comments"
        ]
      },
      "delete": {
        "operationId": "deleteReviewComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ReviewComment ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "ReviewComment deleted."
          },
          "404": {
            "description": "ReviewComment not found."
          }
        },
        "summary": "Delete a ReviewComment.",
        "tags": [
          "review-comments"
        ]
      }
    },
    "/roles": {
      "post": {
        "operationId": "createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Role.",
        "tags": [
          "roles"
        ]
      },
      "get": {
        "operationId": "listRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Role records.",
        "tags": [
          "roles"
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "operationId": "getRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Role not found."
          }
        },
        "summary": "Get a Role by ID.",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Role not found."
          }
        },
        "summary": "Update a Role.",
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role deleted."
          },
          "404": {
            "description": "Role not found."
          }
        },
        "summary": "Delete a Role.",
        "tags": [
          "roles"
        ]
      }
    },
    "/rules": {
      "post": {
        "operationId": "createRule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Rule.",
        "tags": [
          "rules"
        ]
      },
      "get": {
        "operationId": "listRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Rule records.",
        "tags": [
          "rules"
        ]
      }
    },
    "/rules/{id}": {
      "get": {
        "operationId": "getRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Rule not found."
          }
        },
        "summary": "Get a Rule by ID.",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "operationId": "updateRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Rule not found."
          }
        },
        "summary": "Update a Rule.",
        "tags": [
          "rules"
        ]
      },
      "delete": {
        "operationId": "deleteRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Rule deleted."
          },
          "404": {
            "description": "Rule not found."
          }
        },
        "summary": "Delete a Rule.",
        "tags": [
          "rules"
        ]
      }
    },
...
  },
  "info": {
    "title": "IFS Prototype API",
    "description": "OpenAPI contract for the IFS prototype backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RoleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this role.",
            "format": "uuid"
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Role."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Role\" for Role entities.",
            "enum": [
              "Role"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the role."
          },
          "description": {
            "type": "string",
            "description": "Optional human-readable explanation of what this role is for."
          },
          "memberId": {
            "type": "string",
            "description": "Identifier of the member acting through this role."
          },
          "scopeId": {
            "type": "string",
            "description": "Identifier of the scope where this role has authority."
          },
          "scope": {
            "description": "Scope entity describing where this role applies.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScopeResponseDto"
              }
            ]
          },
          "permissions": {
            "description": "Permission entities bundled into this role.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponseDto"
            }
          },
          "state": {
            "enum": [
              "drafted",
              "under-review",
              "active",
              "suspended",
              "revoked",
              "expired"
            ],
            "type": "string",
            "description": "Lifecycle state of the role."
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this role record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this role record was last updated.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "Optional timestamp after which this role no longer applies.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "name",
          "memberId",
          "scopeId",
          "scope",
          "permissions",
          "state",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Comment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Comment\" for Comment entities.",
            "enum": [
              "Comment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this comment was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReviewCommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this review comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this ReviewComment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"ReviewComment\" for ReviewComment entities.",
            "enum": [
              "ReviewComment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this review comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this review comment was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Comment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Comment\" for Comment entities.",
            "enum": [
              "Comment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this comment was last updated.",
            "format": "date-time"
          }
        }
      },
...
      "CreateReviewCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this review comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this ReviewComment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"ReviewComment\" for ReviewComment entities.",
            "enum": [
              "ReviewComment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this review comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this review comment was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateReviewCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this review comment."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this ReviewComment."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"ReviewComment\" for ReviewComment entities.",
            "enum": [
              "ReviewComment"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this review comment was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this review comment was last updated.",
            "format": "date-time"
          }
        }
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this role.",
            "format": "uuid"
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Role."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Role\" for Role entities.",
            "enum": [
              "Role"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the role."
          },
          "description": {
            "type": "string",
            "description": "Optional human-readable explanation of what this role is for."
          },
          "memberId": {
            "type": "string",
            "description": "Identifier of the member acting through this role."
          },
          "scopeId": {
            "type": "string",
            "description": "Identifier of the scope where this role has authority."
          },
          "scope": {
            "description": "Scope entity describing where this role applies.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScopeResponseDto"
              }
            ]
          },
          "permissions": {
            "description": "Permission entities bundled into this role.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponseDto"
            }
          },
          "state": {
            "enum": [
              "drafted",
              "under-review",
              "active",
              "suspended",
              "revoked",
              "expired"
            ],
            "type": "string",
            "description": "Lifecycle state of the role."
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this role record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this role record was last updated.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "Optional timestamp after which this role no longer applies.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "name",
          "memberId",
          "scopeId",
          "scope",
          "permissions",
          "state",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}