# prototype-api/src/review-comment/review-comment.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateReviewCommentDto } from './dto/create-review-comment.dto.js';
import { ReviewCommentResponseDto } from './dto/review-comment-response.dto.js';
import { UpdateReviewCommentDto } from './dto/update-review-comment.dto.js';
import { ReviewCommentService } from './review-comment.service.js';

@ApiTags('review-comments')
@Controller('review-comments')
export class ReviewCommentController {
  constructor(private readonly reviewCommentService: ReviewCommentService) {}

  @Post()
  @ApiOperation({
    operationId: 'createReviewComment',
    summary: 'Create a ReviewComment.',
  })
  @ApiBody({ type: CreateReviewCommentDto })
  @ApiCreatedResponse({ type: ReviewCommentResponseDto })
  create(
    @Body() data: CreateReviewCommentDto,
  ): Promise<ReviewCommentResponseDto> {
    return this.reviewCommentService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listReviewComments',
    summary: 'List ReviewComment records.',
  })
  @ApiOkResponse({ type: ReviewCommentResponseDto, isArray: true })
  findAll(): Promise<ReviewCommentResponseDto[]> {
    return this.reviewCommentService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getReviewComment',
    summary: 'Get a ReviewComment by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ReviewComment ID.' })
  @ApiOkResponse({ type: ReviewCommentResponseDto })
  @ApiNotFoundResponse({ description: 'ReviewComment not found.' })
  async findOne(@Param('id') id: string): Promise<ReviewCommentResponseDto> {
    const result = await this.reviewCommentService.findOne(id);
    if (!result) throw new NotFoundException('ReviewComment not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updateReviewComment',
    summary: 'Update a ReviewComment.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ReviewComment ID.' })
  @ApiBody({ type: UpdateReviewCommentDto })
  @ApiOkResponse({ type: ReviewCommentResponseDto })
  @ApiNotFoundResponse({ description: 'ReviewComment not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateReviewCommentDto,
  ): Promise<ReviewCommentResponseDto> {
    return this.reviewCommentService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deleteReviewComment',
    summary: 'Delete a ReviewComment.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ReviewComment ID.' })
  @ApiNoContentResponse({ description: 'ReviewComment deleted.' })
  @ApiNotFoundResponse({ description: 'ReviewComment not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.reviewCommentService.remove(id);
  }
}