# prototype-api/src/role/role.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateRoleDto } from './dto/create-role.dto.js';
import { RoleResponseDto } from './dto/role-response.dto.js';
import { UpdateRoleDto } from './dto/update-role.dto.js';
import { RoleService } from './role.service.js';

@ApiTags('roles')
@Controller('roles')
export class RoleController {
  constructor(private readonly roleService: RoleService) {}

  @Post()
  @ApiOperation({ operationId: 'createRole', summary: 'Create a Role.' })
  @ApiBody({ type: CreateRoleDto })
  @ApiCreatedResponse({ type: RoleResponseDto })
  create(@Body() data: CreateRoleDto): Promise<RoleResponseDto> {
    return this.roleService.create(data);
  }

  @Get()
  @ApiOperation({ operationId: 'listRoles', summary: 'List Role records.' })
  @ApiOkResponse({ type: RoleResponseDto, isArray: true })
  findAll(): Promise<RoleResponseDto[]> {
    return this.roleService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getRole', summary: 'Get a Role by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Role ID.' })
  @ApiOkResponse({ type: RoleResponseDto })
  @ApiNotFoundResponse({ description: 'Role not found.' })
  async findOne(@Param('id') id: string): Promise<RoleResponseDto> {
    const result = await this.roleService.findOne(id);
    if (!result) throw new NotFoundException('Role not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateRole', summary: 'Update a Role.' })
  @ApiParam({ name: 'id', type: String, description: 'Role ID.' })
  @ApiBody({ type: UpdateRoleDto })
  @ApiOkResponse({ type: RoleResponseDto })
  @ApiNotFoundResponse({ description: 'Role not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateRoleDto,
  ): Promise<RoleResponseDto> {
    return this.roleService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteRole', summary: 'Delete a Role.' })
  @ApiParam({ name: 'id', type: String, description: 'Role ID.' })
  @ApiNoContentResponse({ description: 'Role deleted.' })
  @ApiNotFoundResponse({ description: 'Role not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.roleService.remove(id);
  }
}