# prototype-api/src/rule/rule.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateRuleDto } from './dto/create-rule.dto.js';
import { RuleResponseDto } from './dto/rule-response.dto.js';
import { UpdateRuleDto } from './dto/update-rule.dto.js';
import { RuleService } from './rule.service.js';

@ApiTags('rules')
@Controller('rules')
export class RuleController {
  constructor(private readonly ruleService: RuleService) {}

  @Post()
  @ApiOperation({ operationId: 'createRule', summary: 'Create a Rule.' })
  @ApiBody({ type: CreateRuleDto })
  @ApiCreatedResponse({ type: RuleResponseDto })
  create(@Body() data: CreateRuleDto): Promise<RuleResponseDto> {
    return this.ruleService.create(data);
  }

  @Get()
  @ApiOperation({ operationId: 'listRules', summary: 'List Rule records.' })
  @ApiOkResponse({ type: RuleResponseDto, isArray: true })
  findAll(): Promise<RuleResponseDto[]> {
    return this.ruleService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getRule', summary: 'Get a Rule by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Rule ID.' })
  @ApiOkResponse({ type: RuleResponseDto })
  @ApiNotFoundResponse({ description: 'Rule not found.' })
  async findOne(@Param('id') id: string): Promise<RuleResponseDto> {
    const result = await this.ruleService.findOne(id);
    if (!result) throw new NotFoundException('Rule not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateRule', summary: 'Update a Rule.' })
  @ApiParam({ name: 'id', type: String, description: 'Rule ID.' })
  @ApiBody({ type: UpdateRuleDto })
  @ApiOkResponse({ type: RuleResponseDto })
  @ApiNotFoundResponse({ description: 'Rule not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateRuleDto,
  ): Promise<RuleResponseDto> {
    return this.ruleService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteRule', summary: 'Delete a Rule.' })
  @ApiParam({ name: 'id', type: String, description: 'Rule ID.' })
  @ApiNoContentResponse({ description: 'Rule deleted.' })
  @ApiNotFoundResponse({ description: 'Rule not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.ruleService.remove(id);
  }
}